+++ src/platform.test.ts
@@ -105,7 +105,7 @@ describe('ShellyPlatform', () => {
       matterbridgeDirectory: 'temp',
       matterbridgePluginDirectory: 'temp',
       systemInformation: { ipv4Address: undefined },
-      matterbridgeVersion: '1.6.6',
+      matterbridgeVersion: '1.7.0',
       addBridgedDevice: jest.fn(),
       removeBridgedDevice: jest.fn(),
       removeAllBridgedDevices: jest.fn(),
@@ -324,7 +324,7 @@ describe('ShellyPlatform', () => {
   it('should throw because of version', () => {
     mockMatterbridge.matterbridgeVersion = '1.5.4';
     expect(() => new ShellyPlatform(mockMatterbridge, mockLog, mockConfig)).toThrow();
-    mockMatterbridge.matterbridgeVersion = '1.6.6';
+    mockMatterbridge.matterbridgeVersion = '1.7.0';
   });
 
   it('should call onStart with reason and start mDNS', async () => {